Bugfix `/teleport` command
The `Tier1House0` entry was impossible to match against, because the matched value was forced into all-lowercase before the `switch` block. The match value is now in lowercase as well, allowing it to be matched. Incidentally, the default case has been cleaned up, with the `nelvari6` case moved out from a ternary conditional on the default and into its own case.